move genres below play button row
